Bootstraping API with mocks for every route
Summary:
It was written the first routes of the API and initialized the application.
For now, it will store the users at memory and will not verify if the request is correct.

Test Plan: - Using curl, try GET at /doctors, /doctor/<id> and /patient/<id> and PUT at /doctors and /patients

Reviewers: rpablo007, Bruno

Differential Revision: http://ec2-54-186-131-188.us-west-2.compute.amazonaws.com/D19